Fix errors with multi-line blocks
If a comment to end a multi-line block was instead beginning a new
block, an error would occur due to no indent length being available.

This was a result of the code only checking for indent length at the
start of a multi-line block. For nested blocks, leaving the innermost
block and entering the outermost would result in `indent == None`.

This has been fixed by checking the indent length at both the beginning
and end of a multi-line block. Additionally, the program will prompt the
user to acknowledge any errors as some diff tools will cover the screen
before the error is seen.